requirements.txt
cycler==0.10.0
easydict==1.9
motmetrics==1.2.0
sacred==0.8.0
numpy==1.18.1
torch==1.4.0
matplotlib==3.1.3
opencv_python==4.2.0.32
torchvision==0.5.0
scipy==1.4.1
tqdm==4.42.1
Pillow==7.0.0
PyYAML==5.3
